fix(firestore-bigquery-export): add path_params to multi-thread
